% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_analysis.R
\name{anno_peak_gr37}
\alias{anno_peak_gr37}
\title{Simple wrapper for \code{annotatePeakInBatch}}
\usage{
anno_peak_gr37(
  dba.res,
  annodata = TSS.human.GRCh37,
  output = "nearestBiDirectionalPromoters",
  region = c(-2000, 2000),
  ...
)
}
\arguments{
\item{dba.res}{Any \code{GRanges} object, with gene data, including for instance, a results object from \code{DiffBind}}

\item{annodata}{TSS or other gene data to use for \code{annotatePeakInBatch}. Default = \code{TSS.human.GRCh37}}

\item{output}{Specified mechanism for searching for nearest TSS from interval data. See \code{annotatePeakInBatch} for more.
Default = \code{nearestBiDirectionalPromoter}, which will report TSS's in both directions if the interval overlaps multiple TSS's.}

\item{region}{Annotation range for interval annotation. Default = \code{c(-2000, 2000)} - i.e., 2kb window around the peak}

\item{...}{Any other parameter to be passed to \code{annotatePeakInBatch}}
}
\value{
Annotated GRanges object containing nearest TSS metadata for peaks within 2kb of one
}
\description{
Simple wrapper function for annotatePeakInBatch; will automatically take in a \code{GRanges} object and by default, annotate it
for the gene with the nearest TSS on both strands, assuming the peak falls within +/- 2kb of the TSS.
}
