% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{automater}
\alias{automater}
\title{automater}
\description{
The goal of automater is to automate boring stuff using R with various packages.
}
\author{
Toshikazu Matsumura \href{mailto:matutosi@gmail.com}{matutosi@gmail.com}
}
