% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyautogui.R
\name{center}
\alias{center}
\title{Helper function for recog_image_click() to get center of position.}
\usage{
center(position)
}
\arguments{
\item{position}{A numeric}
}
\value{
A numeric list ($x, $y) of position.
}
\description{
Helper function for recog_image_click() to get center of position.
}
