% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{compareCluster}
\alias{compareCluster}
\title{Yu Lab's compareCluster}
\usage{
compareCluster(
  geneClusters,
  fun = "enrichGO",
  data = "",
  source_from = NULL,
  ...
)
}
\description{
compareCluster function from the Yu Lab (https://github.com/YuLab-SMU/clusterProfiler/blob/master/R/compareCluster.R).
Imported in this package due to versioning issues. See the above link for more information. All credit to Yu et al.
}
