% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDCOMClient.R
\name{convert_docs}
\alias{convert_docs}
\title{Wrapper function to convert document among pdf, xps, html, trf, and txt.
Needs MS Word.}
\usage{
convert_docs(path, format)
}
\arguments{
\item{path}{A string of documents.}

\item{format}{A string of file format to convert.
Avairable: "docx", "pdf", "xps", "html, "rtf", "txt".}
}
\value{
An invisible string of document path.
}
\description{
Wrapper function to convert document among pdf, xps, html, trf, and txt.
Needs MS Word.
}
\examples{
\dontrun{
library(RDCOMClient)
convert_docs("document_path", "pdf")
}
}
