% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_analysis.R
\name{create_samplesheet}
\alias{create_samplesheet}
\title{Creates samplesheet for \code{DiffBind}/\code{ChIPQC}}
\usage{
create_samplesheet(df, bam.dir, peaks.dir, atac = FALSE)
}
\arguments{
\item{df}{Minimal \code{data.frame} object. Assumes that the following columns are present:
\itemize{
\item \code{SampleID} - a unique ID given to each experiment
\item \code{Tissue} - Cell lines or biological samples used in the experiment. These strings should be present in your
\code{.bam} and \code{.xls} filenames.
\item \code{Factor} - the protein/histone mark being ChIPped. This string should be in your \code{.bam} and \code{.xls} filenames.
For ATAC-seq experiments, you may use something like \code{atac-seq}, but this string should also be in the filenames.
\item \code{Condition} - what condition the given experiment is in (e.g., WT, KO, primary tumor, metastasis, etc.)
\item \code{Replicate} - replicate number for the experiment. This number should be present in your read/peak filenames.
}}

\item{bam.dir}{Directory where all \code{.bam} files can be found (should include ALL alignment files!)}

\item{peaks.dir}{Directory where all \code{.xls} peak files from MACS2 can be found (should include ALL peak files!)}

\item{atac}{Boolean indicating whether or not the experiments were ATAC-seq (i.e., have no input controls).
Default = FALSE}
}
\value{
Completed sample sheet that can be immediately used in \code{dba()} from \code{DiffBind} or alternative functions
in \code{ChIPQC}.
}
\description{
Given a minimal \code{data.frame} object, and directories to \code{.bam} and \code{.xls} peak files,
constructs the sample sheet object used for the \code{DiffBind} and \code{ChIPQC} Bioconductor packages.
Note that as of the current development cycle, this function assumes peak calling was performed with
MACS, and that for ChIP-seq experiments, corresponding input controls were used. Please note that
several limiting assumptions are made based on the naming convention of your .bam and .xls files.
This function will automatically match bamReads and bamControls to the appropriate experiment, and
thus requires that naming conventions are uniform for all experiments, and that the user-supplied
information in \code{df} can be found in filenames (when requested - see \code{df} below)
}
\examples{
# Set up data frame for create.samplesheet()
tissue.cai <- rep(c("MDA-MB-231", "MDA-MB-231-BrM2", "MDA-MB-231-LM2"), each = 2)
factor.cai <- rep("h3k27ac", times = length(tissue.cai))
condition.cai <- rep(c("primary", "metastasis", "metastasis"), each = 2)
replicate.cai <- rep(c(1, 2), times = (length(tissue.cai) / 2))
sample.id.cai <- paste0(tissue.cai, replicate.cai)
cai.df <- data.frame(SampleID = sample.id.cai,
                  Tissue = tissue.cai,
                  Factor = factor.cai,
                  Condition = condition.cai,
                  Treatment = treatment.cai,
                  Replicate = replicate.cai)

# Set up directories
cai.bam.dir <- file.path(getwd(), cai.data.dir, "chip-seq/aligned")
cai.peaks.dir <- file.path(getwd(), cai.data.dir, "chip-seq/called_peaks")
# Call samplesheet to use for dba()
cai.samplesheet <- create_samplesheet(cai.df, cai.bam.dir, cai.peaks.dir)
cai.dba <- dba(sampleSheet = cai.samplesheet)
}
