% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_analysis.R
\name{db_analysis}
\alias{db_analysis}
\title{Automate differential binding analysis with \code{DiffBind}}
\usage{
db_analysis(
  sample.df,
  contrast.var,
  summit.val,
  formula.vec = contrast.var,
  contrasts = sample.df[, colnames(contrast.var) == contrast.var] \%>\% unique(),
  blacklist = DBA_BLACKLIST_GRCH37,
  atac = FALSE,
  threshold = 0.05,
  lfc = 0
)
}
\arguments{
\item{sample.df}{Either a minimal \code{data.frame} to be passed to \code{create_samplesheet()} or a sample sheet for use with \code{dba}.
If the \code{bamReads} column is present in \code{sample.df}, the function assumes the user has already set up the full sample sheet.}

\item{contrast.var}{Variable of interest (column in \code{sample.df}) for which DB analysis should be done as a main effect of.}

\item{summit.val}{Integer value to pass for \code{summits} in the \code{dba.count()} function. The \code{summits} parameter in \code{dba.count()} will
be used to set all consensus peaks to a fixed width after counting read pileup and recentering to the point of greatest
read density. The goal of \code{summits}, as described by \code{DiffBind} author Rory Stark, is to take a subsample of the merged peak
that is highly enriched for signal to be used as representative of the entire peak region. This mitigates the effect of background
noise (reads corresponding to nonspecific IP) on the construction of the affinity matrix.}

\item{formula.vec}{Character vector describing the formula design when \code{DiffBind} is passing the affinity matrix to \code{DESeq2}.
Default: equal to \code{contrast.var} (i.e., assumes the user is only interested in the effect of \code{contrast.var} on binding affinity).}

\item{contrasts}{List of character vectors describing all contrasts desired to be run by the user. Each character vector
should be of length 2 and correspond to entries in the column for \code{contrast.var}. Please note, the first-listed contrast
will be in the numerator of the LFC calculation.
Default: will run contrast on all possible unique combinations of the values in \code{contrast.var}}

\item{blacklist}{\code{DiffBind} blacklist object indicating what blacklist to enforce. Set to \code{FALSE} if user does not wish to
enforce any blacklists.
Default: DBA_BLACKLIST_GRCH37}

\item{atac}{Boolean value indicating whether the experiment corresponds to ChIP-seq data (\code{FALSE}) or ATAC-seq data (\code{TRUE}).
If the experiment is ATAC-seq data, the option \code{bUseSummarizeOverlaps} in \code{dba.count} will be set to \code{FALSE} to prevent
memory error crashes.}

\item{threshold}{FDR-adjusted p-value at which to call significantly differentially enriched regions.
Default: 0.05}

\item{lfc}{Real number defining the base LFC at which significantly differentially enriched regions are called.
Default: 0 (i.e., no LFC threshold)}
}
\value{
List of 3 objects:
\itemize{
\item \code{dba.obj}, representing the original DBA object (with any enforced blacklists/greylists)
\item \code{dba.counts}, representing the DBA object containing the count matrix for identified consensus peaks
\item \code{dba.results}, list object containing the output of the \code{dba.analyze} function for each desired contrast.
}
}
\description{
Takes in a samplesheet (or minimal \code{df} to be created into a samplsheet), list of contrasts, and a value for \code{summits}
along with other optional parameters to automate the repeated implementation of the \code{DiffBind} pipeline for differential
binding analysis
}
\examples{
test.analysis <- db_analysis(cai.samplesheet,
                             contrast.var = "Tissue",
                             summit.val = 500,
                             contrasts = list(c("MDA-MB-231-LM2", "MDA-MB-231")))
test.results.list <- test.analysis[[3]]
test.results <- test.results.list[[1]]
test.DB <- dba.report(test.results)
}
