% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_RSelenium.R
\name{deepl_free}
\alias{deepl_free}
\alias{start_selenium}
\alias{stop_selenium}
\alias{sleep}
\title{Wrapper functions to use deepl translation using package RSelenium and others}
\usage{
deepl_free(str, rem, source_lang = "en", target_lang = "ja")

start_selenium()

stop_selenium(rem)

sleep(sec = 5)
}
\arguments{
\item{str}{A character. Maximum length of character is 5000 in deepl_free().}

\item{rem}{An RSelenium object.}

\item{source_lang, target_lang}{A character to select language.
Use capital letters in deepl_api(), and small letters in deepl_free().}

\item{sec}{A numeric (second) to sleep.}
}
\value{
A character translated by deepl
}
\description{
Package deeplr \url{https://cran.r-project.org/web/packages/RSelenium/index.html}
}
\details{
For polite scraping sleep >5 sec in a running of deepl_free().
}
\examples{
library(RSelenium)
library(purrr)
library(dplyr)
library(deeplr)
library(rvest)
\dontrun{
  # Example: from "I am a cat" by Soseki Natsume
sentences <- c("I am a cat. I don't have a name yet.", "I have no idea where I was born.")
api_key <- "your_api_key"
tibble::tibble(en = sentences) \%>\%
  dplyr::mutate(jp = deepl_api(en, api_key = api_key))
rem <- start_selenium()
purrr::map_chr(sentences, deepl_free, rem = rem)
stop_selenium(rem)
}
}
