% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyautogui.R
\name{display_corner}
\alias{display_corner}
\title{Helper function for recog_image_click().}
\usage{
display_corner(corner = "bottom_right", width = 600, height = 600)
}
\arguments{
\item{corner}{A string to specify a corner.
"top_left", "top_right", "bottom_left", or "bottom_right".}

\item{width, height}{A integer.}
}
\description{
Helper function for recog_image_click().
}
