% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyautogui.R
\name{display_size}
\alias{display_size}
\title{Helper function for display_corner() to get display size.}
\usage{
display_size()
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A numeric list ($width, $height) of display size.
}\if{html}{\out{</div>}}
}
\description{
Helper function for display_corner() to get display size.
}
