% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDCOMClient.R
\name{docx2pdf}
\alias{docx2pdf}
\title{Wrapper function to convert docx into pdf using package RDCOMClient.}
\usage{
docx2pdf(path)
}
\arguments{
\item{path}{A string of docx path to convert.}
}
\value{
An invisible string of pdf path.
}
\description{
Wrapper function to convert docx into pdf using package RDCOMClient.
}
\examples{
\dontrun{
docx2pdf()
}

}
