% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{dotplot_compareClusterResult_gsea}
\alias{dotplot_compareClusterResult_gsea}
\title{Modified dotplot for \code{compareClusterResult} objects.}
\usage{
dotplot_compareClusterResult_gsea(
  object,
  x = "Cluster",
  colorBy = "NES",
  showCategory = 5,
  by = "geneRatio",
  size = "p.adjust",
  split = NULL,
  includeAll = TRUE,
  font.size = 12,
  title = "",
  label_format = 30,
  group = FALSE,
  shape = FALSE
)
}
\description{
Dotplot modified from the \code{clusterProfiler} package to be more amenable to viewing
GSEA results from the \code{compareCluster} package. Instead of the doplot being colored by \code{p-adjust} and
sized by \code{GeneRatio}, the plot is colored by \code{NES} and sized by \code{p-adjust} for visibility in which
processes are enriched vs. depleted in each phenotype. See https://github.com/YuLab-SMU/enrichplot/blob/master/R/dotplot.R
for the original code. All credit to Yu et al.
}
