% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{dotplot_enrichResult_Col}
\alias{dotplot_enrichResult_Col}
\title{Modified dotplot for \code{gseaResult} objects}
\usage{
dotplot_enrichResult_Col(
  object,
  x = "NES",
  color = "p.adjust",
  showCategory = 10,
  size = NULL,
  split = NULL,
  font.size = 12,
  title = "",
  orderBy = "x",
  label_format = 30,
  decreasing = TRUE
)
}
\description{
Dotplot modified from Yu et al.'s \verb{clusterProfiler' to made more amenable to GSEA results. Instead of plotting }GeneRatio\verb{, the }NES\verb{is plotted as a horizontal bar plot and colored by}p-adjusted`.
See https://github.com/YuLab-SMU/enrichplot/blob/master/R/dotplot.R for the original code. All credit
to Yu et al.
}
