% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{double_quote}
\alias{double_quote}
\title{Add double quotation mark before and after string}
\usage{
double_quote(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string.
}
\description{
Add double quotation mark before and after string
}
\examples{
double_quote("some string")

}
