% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_analysis.R
\name{ensembl_to_gene}
\alias{ensembl_to_gene}
\title{Ensembl ID conversion}
\usage{
ensembl_to_gene(ensembl.genes)
}
\arguments{
\item{ensembl.genes}{A character vector of Ensembl IDs.}
}
\value{
A character vector containing the matching HGNC symbol for each Ensembl ID.
NA indicates the symbol could not be matched to a gene feature.
}
\description{
Takes a vector of Ensembl IDs and converts them to the corresponding HGNC gene symbols using the \code{EnsDb.Hsapiens.v79} package.
}
\examples{
ens.vec <- c("ENSG00000166035", "ENSG00000102554", "ENSG00000141510")
ensembl_to_gene(ens.vec)
}
