% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exclude}
\alias{exclude}
\title{Exclude vector elements that matches a condition.}
\usage{
exclude(x, cond)
}
\arguments{
\item{x}{A vector.}

\item{cond}{A logical.}
}
\value{
A vector.
}
\description{
Exclude vector elements that matches a condition.
}
\examples{
x <- 1:10
x[! x < 5 ]
x \%>\%
  exclude(x < 5)

}
