% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_officer.R
\name{extract_text}
\alias{extract_text}
\title{Wrapper functions to extract text using package officer.}
\usage{
extract_text(doc, drop_na = TRUE, drop_empty = TRUE)
}
\arguments{
\item{doc}{A doc object.}

\item{drop_na, drop_empty}{A logical.}
}
\description{
Wrapper functions to extract text using package officer.
}
