% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_numbered}
\alias{file_numbered}
\title{Add numbers to file name.}
\usage{
file_numbered(file, n, extra = 0, sep = "_")
}
\arguments{
\item{file}{A string of file name.}

\item{n, extra}{An integer.}

\item{sep}{A string of separator between file name and number.}
}
\value{
String vector of numbered file name.
}
\description{
Add numbers to file name.
}
\examples{
file <- "aaa.body.ext"
file_numbered(file, 30)

}
