% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_split_name}
\alias{file_split_name}
\title{Split file name into body part and extention.}
\usage{
file_split_name(file)
}
\arguments{
\item{file}{A string of file name.}
}
\value{
A list including file name of body part and extention.
}
\description{
Split file name into body part and extention.
}
\examples{
file <- "aaa.body.ext"
file_split_name(file)

}
