% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{find_python}
\alias{find_python}
\title{Find a path of Python.}
\usage{
find_python(select_memu = TRUE)
}
\arguments{
\item{select_menu}{A logical to show select menu.
TRUE: show select menu and return one path
FALSE: return all found paths.}
}
\value{
A string or string vector.
}
\description{
Find a path of Python.
}
\examples{
python_path <- find_python()
reticulate::use_python(python_path)

}
