% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{format_ymd}
\alias{format_ymd}
\alias{extract_date}
\alias{remove_empty_chr}
\alias{add_year}
\title{Wrapper functions to ... using package lubridate}
\usage{
format_ymd(x, yr = "")

extract_date(x)

remove_empty_chr(x)

add_year(x, yr)
}
\arguments{
\item{x, yr}{String(s) of date and year.}
}
\value{
(A list of) strings.
}
\description{
Package lubridate \url{https://cran.r-project.org/web/packages/lubridate/index.html}
}
\examples{
x <- 
  "21年1月1日(月)，2021年1月1日(月)，2021年10月10日(月)，
  2月2日(月)，12月22日(月)，2021/1/1(月)，2021/10/10(月)，
  2/2(月)，12/22(月)，21年1月1日，2021年1月1日，
  2021年10月10日，2月2日，12月22日，2021/1/1，2021/10/10，
  2/2，12/22"
format_ymd(x)

}
