% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{index2xy}
\alias{index2xy}
\title{Convert array index into xy location in matrix.
Helper function for locate_ndl_in_hay().}
\usage{
index2xy(index, nrow)
}
\arguments{
\item{index, nrow}{A numeric.}
}
\value{
A pair numeric of xy postion.
}
\description{
Convert array index into xy location in matrix.
Helper function for locate_ndl_in_hay().
}
\examples{
nrow <- 4
matrix(1:12, nrow = nrow)
purrr::map(1:12, index2xy, nrow = nrow)

}
