% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_duplicated}
\alias{is_duplicated}
\title{Add numbers to file name.}
\usage{
is_duplicated(x, y)
}
\arguments{
\item{x, y}{A string vector.}
}
\value{
A logical.
}
\description{
Add numbers to file name.
}
\examples{
x <- letters[1:5]
y <- letters[5:8]
z <- letters[6:8]
is_duplicated(x, y)
is_duplicated(x, z)
is_duplicated(y, z)

}
