% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{locate_image_r}
\alias{locate_image_r}
\title{Locate image position}
\usage{
locate_image_r(needle_image, haystack_image, center = TRUE)
}
\arguments{
\item{needle_image, haystack_image}{A image.}

\item{center}{A logical.
TRUE returns center position of needle_image.}
}
\value{
A vector (x-y) integer of location for ndl_img.
}
\description{
Locate image position
}
\examples{
library(magrittr)
library(imager)
haystack_image <- imager::load.example("parrots")
pos_x <- 200; pos_y <- 200
size_x <- 50; size_y <- 20
needle_image <- 
  haystack_image[
    pos_x:(pos_x + size_x), 
    pos_y:(pos_y + size_y),] \%>\%
  as.cimg()

layout(t(1:2))
plot(needle_image)
plot(haystack_image)

locate_image(needle_image, haystack_image)

}
