% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ls_class}
\alias{ls_class}
\title{Add double quotation mark before and after string}
\usage{
ls_class()
}
\value{
A list of class name with its object name.
}
\description{
reference: https://search.r-project.org/CRAN/refmans/FFD/html/lls.html
}
\examples{
ls_class()

}
