% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick.R
\name{magick_crop}
\alias{magick_crop}
\title{Wrapper functions to convert images using package magick}
\usage{
magick_crop(image, left_top, right_bottom)
}
\arguments{
\item{image}{A magick-image object.}

\item{left_top, right_bottom}{A numeric vector of x- and y-position.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                  A magick-image object.
}\if{html}{\out{</div>}}
}
\description{
Package magick \url{https://cran.r-project.org/web/packages/magick/index.html}
}
\examples{
\donttest{
library(magick)
logo <- image_read("logo:")
magick_crop(logo, c(100,100), c(200,250))
}

}
