% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{make_shortcut}
\alias{make_shortcut}
\alias{add_path}
\title{Make shortcut using WsShell on Windows}
\usage{
make_shortcut(
  exe,
  shortcut = NULL,
  dir = NULL,
  arg = NULL,
  size = 1,
  wd = NULL
)

add_path(new_path)
}
\arguments{
\item{exe, shortcut, dir, arg, wd}{A string of exe file, shortcut name, directory of shortcut,
command line arguments, and working directory.}

\item{new_path}{A string of path to add.}
}
\value{
size  A numeric for window size. 1: normal, 3: max, 7: min.

make_shortcut() returns A list including shortcut path, and result of shell(),
add_path() returns A result of shell().
}
\description{
See detail in the URL below.
https://learn.microsoft.com/troubleshoot/windows-client/admin-development/create-desktop-shortcut-with-wsh
}
\examples{
library(magrittr)
exe <- fs::path(Sys.getenv("R_HOME"), "bin/x64/Rgui.exe")
shortcut <- "r"
arg <- "--no-restore --no-save --sdi --silent"
wd <- Sys.getenv("R_USER")
make_shortcut(exe, shortcut = shortcut, arg = arg, wd = wd)

}
