% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_to_continue}
\alias{message_to_continue}
\title{Display message to continue}
\usage{
message_to_continue(msg = "Press [Enter] key to continue ...")
}
\arguments{
\item{msg}{A string to show.}
}
\description{
Display message to continue
}
