% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{mouse_record}
\alias{mouse_record}
\title{Wrapper function for recording mouse.get_cursor() position}
\usage{
mouse_record(n = 5, interval = 1)
}
\arguments{
\item{n, interval}{A numeric of record number and interval (sec).
"interval < 0" means wait user input (any keys on R console).}
}
\value{
A list of x and y position.
}
\description{
Wrapper function for recording mouse.get_cursor() position
}
\examples{
mouse_record(interval = 0.1)

}
