% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlook_Microsoft365R.R
\name{outlook}
\alias{outlook}
\alias{create_email}
\alias{gen_email}
\alias{send_email}
\alias{add_attachment}
\title{Wrapper function to create and send outlook emails based on the contents of excel file using package Microsoft365R.}
\usage{
create_email(path, outlook)

gen_email(outlook, df)

send_email(emails)

add_attachment(emails, df)
}
\arguments{
\item{path}{A string of path for MS excel that include columns of email:
"send" (required),
1: send, 0: save in drafts.
"to" (required).
"subject" (almost required), "body" (almost required).
"cc" (optional), "bcc" (optional).
"attachment" (optional).
"attachment" is path to files separated with comma (",").}

\item{outlook}{ms_outlook_object generated with get_business_outlook().}

\item{df}{A dataframe including contents of emails.}

\item{emails}{A list of ms_outlook_email class.}
}
\value{
A list of ms_outlook_email class.
}
\description{
Package Microsoft365R \url{https://cran.r-project.org/web/packages/Microsoft365R/index.html}
}
\examples{
\dontrun{
outlook <- Microsoft365R::get_business_outlook()
path <- "outlook_Microsoft365R.xlsx"
create_email(path, outlook)
}

}
