% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_convert}
\alias{path_convert}
\title{Convert path strings.}
\usage{
path_convert(path, sub_dir = NULL, pre = NULL, post = NULL, ext = NULL)
}
\arguments{
\item{path, sub_dir, pre, post, ext}{A string of path, sub directory, prepositive, postpositive, and extension.}
}
\value{
A string of path.
}
\description{
Convert path strings.
}
\examples{
library(tidyverse)
fs::path_package("automater", "rsc") \%>\%
  fs::dir_ls() \%>\%
  print() \%>\%
  path_convert(sub_dir = "sub", pre = "PRE_", post = "_POST", ext = "R")

}
