% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg      A string of package names.}
\alias{pkg      A string of package names.}
\alias{validate_package}
\title{Check package installation and install it when not yet.}
\usage{
validate_package(pkg, repos = "https://cran.ism.ac.jp/")
}
\arguments{
\item{reposs}{A string of url for repository}
}
\description{
Check package installation and install it when not yet.
}
