% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyautogui.R
\name{recog_image_click}
\alias{recog_image_click}
\title{Wrapper function for PyAutoGUI to recognize and click image.}
\usage{
recog_image_click(img, pag, wait = TRUE, button = "left", hold = FALSE, ...)
}
\arguments{
\item{img}{A string of path for image to recognize.}

\item{pag}{A python.builtin.module built
by reticulate::import("pyautogui")}

\item{wait}{A logical.
TRUE: Wait untill finding a image.
FALSE: When not found image, return error.}

\item{button}{A string of mouse: "left" or "right".}

\item{hold}{A logical.
TRUE: Keep holding.}

\item{...}{More arguments to pass pag$locateOnScreen().}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A invisible list ($x, $y) of clicked position.
}\if{html}{\out{</div>}}
}
\description{
Wrapper function for PyAutoGUI to recognize and click image.
}
\examples{
python_path <- find_python()
reticulate::use_python(python_path)
pag <- reticulate::import("pyautogui")
img <- fs::path_package("automater", "img", "up_arrow.png")
recog_image_click(img, pag)

}
