% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_officer.R
\name{replace_doc_officer}
\alias{replace_doc_officer}
\title{Wrapper functions to replace text in MS Words using package officer.}
\usage{
replace_doc_officer(
  doc,
  old_value,
  new_value,
  body = TRUE,
  header = TRUE,
  footer = TRUE
)
}
\arguments{
\item{doc}{A doc object.}

\item{old_value}{A string of old value.}

\item{new_value}{A string of new value.}

\item{body, header, footer}{A logical to be replaced or not.}
}
\value{
replace_doc_officer() returns a doc object.
}
\description{
Package qpdf \url{https://cran.r-project.org/web/packages/officer/index.html}
}
\examples{
\dontrun{
  #   doc <- officer::read_docx(path)

}

}
