% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_officer.R
\name{replace_docs}
\alias{replace_docs}
\title{Helper function}
\usage{
replace_docs(path, replacement)
}
\arguments{
\item{path, replacement}{A string of path or replacement.}
}
\description{
Helper function
}
