% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_gsea_cp}
\alias{run_gsea_cp}
\title{Run \code{clusterProfiler} GSEA}
\usage{
run_gsea_cp(ranked.list, category, ...)
}
\arguments{
\item{ranked.list}{Ranked list in the form of a named numeric vector, sorted in descending order with Entrez ID identifiers.}

\item{category}{Category from MSigDB to use by \code{GSEA()}. Will be fetched using \code{msigdbr()}.}

\item{...}{Additional arguments to pass to \code{GSEA()}.}
}
\value{
\code{gseaResult} object containing the results of the call to \code{GSEA()}.
}
\description{
Simple wrapper function for the \code{GSEA} function in the \code{clusterProfiler} package. Utilizes \code{msigdbr} to read in
a hallmark set and run \code{GSEA} with a pre-ranked list (which can be constructed with the \code{prepare_gsea()} function).
}
