% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.R
\name{set_af_fp2}
\alias{set_af_fp2}
\title{Wrapper functions to manipulate excel work books and sheets using package xlsx.}
\usage{
set_af_fp2(file)
}
\arguments{
\item{file}{A string of excel file name.}
}
\description{
Package xlsx \url{https://cran.r-project.org/web/packages/openxlsx/index.html}
}
\examples{
\donttest{
files <- list.files(pattern = "xls")
for(file in files){
  automater::set_af_fp(file)
}
}

}
