% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_rsc}
\alias{set_rsc}
\title{Copy *.rsc (and *.command, and chmod on Mac)}
\usage{
set_rsc(file, path)
}
\arguments{
\item{file, path}{A string of file and path}
}
\description{
Copy *.rsc (and *.command, and chmod on Mac)
}
