% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{shot}
\alias{shot}
\title{Take a screenshot.
Need to download screenCapture.exe or to compile from screenCapture.bat.}
\usage{
shot(file = "")
}
\arguments{
\item{file}{A string for file name of screenshot.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A file name of screenshot. When "", screenshot will be saved in a tempral directory.
}\if{html}{\out{</div>}}
}
\description{
"screenCapture.bat"  https://github.com/npocmaka/batch.scripts/blob/master/hybrids/.net/c/screenCapture.bat
}
\examples{
\dontrun{
library(imager)
sc <- screenshot()
imager::load.image(sc)
}

}
