% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sleep}
\alias{sleep}
\title{Wrapper function for Sys.sleep()}
\usage{
sleep(sec = 5)
}
\arguments{
\item{sec}{A numeric (second) to sleep.}
}
\value{
Invisible NULL.
}
\description{
Wrapper function for Sys.sleep()
}
\examples{
sleep(0.1)

}
