% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_RSelenium.R
\name{sleep_rnd}
\alias{sleep_rnd}
\title{Wrapper function for Sys.sleep()}
\usage{
sleep_rnd(sec = 5)
}
\arguments{
\item{sec}{A numeric (second) to sleep.}
}
\description{
To avoid constant sleep, random second is given by runif() and rnorm().
}
