% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_analysis.R
\name{splitDE}
\alias{splitDE}
\title{Isolate up- and down-regulated genes}
\usage{
splitDE(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{signifDE} (see \code{?signifDE} for more)}
}
\value{
A list of 2 objects, containing 1) \code{results} subset that has significantly upregulated genes,
and 2) a \code{results} subset that has significantly downregulated genes.
}
\description{
Splits a \verb{DESeq2 results} object into genes that are upregulated in the contrast (LFC > 0) and
genes that are downregulated in the contrast (LFC < 0). Makes an internal call to \code{signifDE()}
such that only genes whose change in regulation is statistically significant is returned.
}
