% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_RSelenium.R
\name{start_selenium}
\alias{start_selenium}
\title{Helper function to start Selenium and RSelenium
Need instration of Java, Selenium and GoogleChrome driver as described in README.R}
\usage{
start_selenium()
}
\description{
Helper function to start Selenium and RSelenium
Need instration of Java, Selenium and GoogleChrome driver as described in README.R
}
