% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_RSelenium.R
\name{stop_selenium}
\alias{stop_selenium}
\title{Helper function to stop Selenium and RSelenium}
\usage{
stop_selenium(rem)
}
\arguments{
\item{rem}{An RSelenium object.}
}
\description{
Helper function to stop Selenium and RSelenium
}
