% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_analysis.R
\name{txt2counts}
\alias{txt2counts}
\title{Reading in raw RNA-seq data}
\usage{
txt2counts(featurecounts.txt)
}
\arguments{
\item{featurecounts.txt}{Path to valid \code{featureCounts} filename}
}
\value{
A data frame containing columns corresponding to samples and rows corresponding to raw gene counts.
}
\description{
Converts a featureCounts count file into a data frame for downstream analysis.
}
\examples{
# Reading in a featureCounts file from directory "~/counts"
# featurecounts.txt <- file.path("~/counts", "counts.txt")
# txt2counts(featurecounts.txt)
}
