% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_RSelenium.R
\name{use_deepl}
\alias{use_deepl}
\alias{deepl_api}
\alias{deepl_free}
\title{Wrapper functions to use deepl translation using package deeplr}
\usage{
deepl_api(str, api_key, source_lang = "EN", target_lang = "JA")

deepl_free(str, rem, source_lang = "en", target_lang = "ja")
}
\arguments{
\item{str}{A character. Maximum lenght of character is 5000 in deepl_free().}

\item{api_key}{A character of deepl api key}

\item{source_lang, target_lang}{A character to select language.
Use capital letters in deepl_api(), and small letters in deepl_free().}

\item{rem}{An RSelenium object.}
}
\value{
A character translated by deepl
}
\description{
Package deeplr \url{https://cran.r-project.org/web/packages/deeplr/index.html}
}
\details{
Use deepl by api (deepl_api()) and free (deepl_free())
For polite scraping sleep >5 sec in a running of deepl_free().
}
\examples{
library(RSelenium)
library(tidyverse)
library(deeplr)
library(rvest)

  # Example: "I am a cat" by Soseki Natsume
sentences <- c("I am a cat. I don't have a name yet.", "I have no idea where I was born.")
api_key <- "ce943af7-40e7-a8bc-a44b-466e15ee1059:fx"
tibble::tibble(en = sentences) \%>\%
  dplyr::mutate(jp = deepl_api(en, api_key = api_key))
rem <- start_selenium()
purrr::map_chr(sentences, deepl_free, rem = rem)
stop_selenium(rem)


  # Example: Tidy modelong with R (https://www.tmwr.org/)

api_key <- "ce943af7-40e7-a8bc-a44b-466e15ee1059:fx"
tidy_modeling <- readr::read_tsv("D:/matu/work/ToDo/TMwR/jp/contens.txt")
str <- 
  tidy_modeling \%>\%
  `[[`("en")
str \%>\%
  purrr::map_dbl(stringr::str_length) \%>\%
  max()

rem <- start_selenium()

str <- head(str, 9)
n <- 500
n <- 5
len <- length(str)
times <- (len \%/\% n) + ifelse((len \%\% n) == 0, 0, 1)
jp <- list()

for(i in seq(times)){
  begin <- (i - 1) * n + 1
  end   <- min(i * n, len)
  tmp_str <- str[begin:end]
  jp[[i]] <- purrr::map_chr(tmp_str, deepl_free, rem = rem)
  print(paste0(i * n, " / ", len))
}

stop_selenium(rem)
jp <- unlist(jp)
readr::write_tsv(tibble::tibble(jp), "C:/Users/matsumura/Desktop/tmwr_jp.txt")



}
