% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_input}
\alias{user_input}
\title{Show message on prompt and Receieve user input from stdin}
\usage{
user_input(prompt)
}
\arguments{
\item{prompt}{A string message to show on prompt}
}
\value{
A string of user input
}
\description{
Show message on prompt and Receieve user input from stdin
}
