% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_package}
\alias{validate_package}
\title{Check package installation and install it when not yet.}
\usage{
validate_package(pkg, repos = "https://cran.ism.ac.jp/")
}
\arguments{
\item{pkg}{A string of package names.}

\item{repos}{A string of url for repository}
}
\description{
Check package installation and install it when not yet.
}
\examples{
\donttest{
validate_package("automater")
validate_package("xlsx")
}

}
