% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_analysis.R
\name{volcano.plot}
\alias{volcano.plot}
\title{Volcano plot}
\usage{
volcano.plot(de.results, lfc = 0, threshold = 0.05)
}
\arguments{
\item{de.results}{Object produced by a call to DESeq2 \code{results}.}

\item{lfc}{Threshold for determining significant change in gene expression.
E.g., lfc = 1 indicates only genes with a p-adjusted < \code{threshold} and \code{abs(LFC)} > \code{lfc} should be deemed
significantly different in gene expression across the comparison samples. Default = 0 (i.e., no LFC threshold)}

\item{threshold}{FDR-adjusted p-value cutoff for statistical significance. Default = 0.05.}
}
\value{
Plot object produced by a call to \code{ggplot} from the \code{ggplot2} package.
}
\description{
Produces a volcano plot displaying -log10(p-adj) vs. LFC for genes given a \verb{DESeq2 results} object.
}
