% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick.R
\name{web_screenshot}
\alias{web_screenshot}
\title{Donwload screenshot image from web based on URL and
add text annotation of URL on the top of the image.}
\usage{
web_screenshot(
  url,
  trim = TRUE,
  border_size = "x40",
  annotate_size = 20,
  format = "png",
  resize = FALSE
)
}
\arguments{
\item{url}{A string of URL.}

\item{trim}{A logical if trim image by magick::image_trim().}

\item{border_size}{A string of border size. "x40": Set border
on the top for annotation. Will be passed like
imagick::image_border(img, "white", geometry = border_size).}

\item{annotate_size}{A string of annnotation size.
Will be passed like magick::image_annotate
(img, url, size = annotate_size).}

\item{format}{A string of format of image. Will be passed like
magick::image_convert(img, format = format)}

\item{resize}{A string for resize by magick::image_scale().
Passed like magick::image_scale(geometry = resize).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         A magick-image object.
}\if{html}{\out{</div>}}
}
\description{
Donwload screenshot image from web based on URL and
add text annotation of URL on the top of the image.
}
\examples{
\donttest{
url <- "https://www.deepl.com/translator"
img <- web_screenshot(url, format = "png")
magick::image_browse(img)
magick::image_write(img, "deepl.png")
}

}
