% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_officer.R
\name{write_docx}
\alias{write_docx}
\title{Wrapper functions to write docx using package officer.}
\usage{
write_docx(doc, path)
}
\arguments{
\item{doc}{A doc object.}

\item{path}{A string of path.}
}
\description{
Wrapper functions to write docx using package officer.
}
