% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{xlsx}
\alias{xlsx}
\alias{set_auto_filter}
\title{Wrapper function to manipulate excel work books and sheets.}
\usage{
set_auto_filter(sh)
}
\arguments{
\item{sh}{A excel sheet.}

\item{file}{A string of excel file name.}
}
\description{
Package xlsx <>
incudes usefull functions as shown bellow.
}
