% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{xy_pos}
\alias{xy_pos}
\title{Get xy position of a value in a matrix
Helper function for locate_ndl_in_hay()}
\usage{
xy_pos(mt, val)
}
\arguments{
\item{mt}{A matrix}

\item{val}{A matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list of xy pairs.
}\if{html}{\out{</div>}}
}
\description{
Get xy position of a value in a matrix
Helper function for locate_ndl_in_hay()
}
\examples{
mt <- matrix(1:12, nrow = nrow)
xy_pos(mt, 5)

}
