% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{yr_regexp}
\alias{yr_regexp}
\title{unescape_utf() year, month, day and day of the week
for regular expressions of date}
\usage{
yr_regexp()
}
\description{
unescape_utf() year, month, day and day of the week
for regular expressions of date
}
