% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/keyboardsimulator.R
\name{is_equal_matrix}
\alias{is_equal_matrix}
\alias{locate_image_cpp}
\alias{image2gray_matrix}
\alias{locate_image}
\title{Search needle image in haystack image and locate needle position.}
\usage{
is_equal_matrix(mt_1, mt_2)

locate_image_cpp(ndl_mt, hay_mt)

image2gray_matrix(img)

locate_image(needle_image, haystack_image, center = TRUE)
}
\arguments{
\item{mt_1, mt_2}{A matrix of image data.}

\item{ndl_mt, hay_mt}{A matrix with grayscale x-y position.}

\item{img}{A Cimg.}

\item{needle_image, haystack_image}{A image.}

\item{center}{A logical.
TRUE returns center position of needle_image.}
}
\value{
A logical.

x-y position of needle image.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ A two (x-y) dimensional matrix.
}\if{html}{\out{</div>}}

A vector (x-y) integer of location for ndl_img.
}
\description{
Search needle image in haystack image and locate needle position.

Combert Cimg class into grayscale x-y matrix.
Use grayscale to Speed up and to simplify code.
}
\examples{
library(magrittr)
library(imager)
haystack_image <- imager::load.example("parrots")
pos_x <- 200; pos_y <- 200
size_x <- 50; size_y <- 20
needle_image <- 
  haystack_image[
    pos_x:(pos_x + size_x), 
    pos_y:(pos_y + size_y),] \%>\%
  as.cimg()

layout(t(1:2))
plot(needle_image)
plot(haystack_image)

locate_image(needle_image, haystack_image)

}
