% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_equal_matrix}
\alias{is_equal_matrix}
\alias{locate_image_cpp}
\title{Search needle image in haystack image and locate needle position.}
\usage{
is_equal_matrix(mt_1, mt_2)

locate_image_cpp(ndl_mt, hay_mt)
}
\arguments{
\item{mt_1, mt_2}{A matrix of image data.}

\item{ndl_mt, hay_mt}{A matrix with grayscale x-y position.}
}
\value{
A logical.

x-y position of needle image.
}
\description{
Search needle image in haystack image and locate needle position.
}
