% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardsimulator.R
\name{mouse_move_click}
\alias{mouse_move_click}
\title{Wrapper function for mouse.move() and mouse.click()}
\usage{
mouse_move_click(x, y)
}
\arguments{
\item{x, y}{A numeric to move.}
}
\value{
Invisible NULL.
}
\description{
Wrapper function for mouse.move() and mouse.click()
}
\examples{
mouse_move_click(100, 100)

}
