% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_officer.R
\name{replace_docs}
\alias{replace_docs}
\title{Helper function}
\usage{
replace_docs(replacement)
}
\arguments{
\item{replacement}{A daraframe including cols: "file", "old_value", and "new_value".}
}
\description{
Helper function
}
